requirements.txt
python-telegram-bot==22.4
sqlalchemy==2.0.43
python-dotenv==1.1.1
APScheduler==3.11.0
tzlocal==5.3.1
jdatetime==5.2.0
pillow==11.3.0
reportlab==4.4.4
requests==2.32.5
